Better error messages for api_version mismatch. Also be a bit stricter about when to set the error, and the types of version numbers.

+++ include/cwipc_codec/api.h
@@ -40,8 +40,8 @@ struct cwipc_encoder_params
  * also pass this version number, to ensure your code is compatible with this version
  * of the library.
  */
-#define CWIPC_ENCODER_PARAM_VERSION 0x20220607
-#define CWIPC_ENCODER_PARAM_VERSION_OLD 0x20190506
+#define CWIPC_ENCODER_PARAM_VERSION ((uint32_t)0x20220607)
+#define CWIPC_ENCODER_PARAM_VERSION_OLD ((uint32_t)0x20190506)
 
 #ifdef __cplusplus
